Do sharing analysis with constraint generation
This works around an aliassing issue, e.g in

    (\val -> val) (\val -> val)

Or

    case b when
        Foo x -> x + x
        Bar x -> x

In both cases, the two functions/two branches have a variable with the
same symbol, but one that is really distinct.

Thus after processing a function or case or let (not added yet), the
bound variables must be removed from the var_usage map.