properly generate lets
The topological sort into groups only give groups whose elements are equivalent with regards to the topological sort. But there may still be internal cycles withing a group, or the group can contain no mutual recursive definitions at all! So we must look more carefully at the groups, using the strongly connected components.

So that is what happens now: The normal analysis uses a modified successor function that is syntactic: it gives all symbols in a definition's body, even (especially!) those behind a lambda.
With this, we can separate recursive functions from non-recursive ones.

Then a second analysis is needed to crudely spot invalid recursion, of the kind `x = x`. Here, we don't want to look behind a lambda.

I also had to change some test cases to actually use a symbol in the result: this analysis doesn't run for unused symbols.