A rough first version
When hitting an Access node, this will see if the chain ends in a Var
node. If so, the chain of fields and the var are joined into one new
symbol, that is registered as a used variable.

Now what needs to happen is that the usage of r.x influences the type of
that field.