Implement constraint generation for annotated `Expr2::FunctionDef`s
This is mostly the same as the procedure for unannotated `FunctionDef`s.
The only differences are that we need to

1. note the rigid type variables that are a part of the FunctionDef
2. generate fresh type variables for the function arguments, and
   constrain them to the types of the arguments passed to us in the
   annotation

This is enough to get the constraint generation working. There are more
interesting cases that properly typecheck in the main compiler but do
not typecheck given the current state of the editor AST (see the second
test in this commit), but this is a bug due to canonicalization, not
type reconstruction.