www: Change download URL for pre-built web REPL
I'm making changes more frequently to the REPL than we are to the source code.
I want to be able to deploy it from a script rather than opening new GitHub issues,
so I decided to just commit the assets into a branch of a GitHub repo.
The repo was only used as a prototype for the web REPL so I don't use it anymore.