Unhelpful compiler error (wrong location with invalid `href=`
── UNFINISHED LIST ────────────────── examples/static-site-gen/static-site.roc ─

I am partway through started parsing a list, but I got stuck here:

21│          head [] [
22│              meta [httpEquiv "content-type" , content "text/html; charset=utf-8" ] [],
23│              title [] [text "Daring Fireball: Markdown"],
24│              link [rel "apple-touch-icon-precomposed" ,href="/graphics/apple-touch-icon.png"] [],
                 ^

I was expecting to see a closing square bracket before this, so try
adding a ] and see if that helps?

Note: When I get stuck like this, it usually means that there is a
missing parenthesis or bracket somewhere earlier. It could also be a
stray keyword or operator.